src: lib: drivers: Extract mavlink json driver into its own file
